Stabilize proc/debug ABI formatting and improve debug shell UX
Replace unsupported printf length modifiers across proc/debug paths to prevent
literal format artifacts and zero-valued parsing in userspace-like debug bins.
This fixes incorrect runtime outputs in commands such as free and uptime and
hardens text ABI generation for 64-bit counters.

Key changes:
- Add decimal u64 formatting helpers that do not rely on %lu/%llu support.
- Rework /proc producers to emit portable numeric text for:
  - meminfo
  - uptime
  - blockdevices
  - partitions
  - iomem
  - timer_list
- Remove non-portable float/size_t formatting in /proc/loadavg and emit
  fixed-point values with integer-safe formatting.
- Update debug bins parsers (free/uptime/lsblk) to skip non-digit prefixes
  before numeric parsing for resilience.
- Improve free output readability with explicit kB header labels.
- Improve lsblk presentation:
  - clearer column layout (NAME/TYPE/MAJ:MIN/SIZE/MODEL)
  - better alignment for disks/partitions
  - conditional serial rendering only when available
- Add tab-completion to dbgshell for first-token command names:
  - single match: complete and append trailing space
  - multiple matches: expand common prefix
  - ambiguous set: print candidates and redraw prompt/input
- Keep debug bins syscall-only behavior intact (no kernel-internal coupling).
- Preserve build stability: compile/link/ISO generation validated after changes.

+++ debug_bins/cmd_date.c
@@ -64,7 +64,7 @@ static int cmd_date_handler(int argc, char **argv)
     char line[80];
     format_timestamp((time_t)tv.tv_sec, line, sizeof(line));
     size_t len = strlen(line);
-    snprintf(line + len, sizeof(line) - len, ".%06ld UTC\n", (long)tv.tv_usec);
+    snprintf(line + len, sizeof(line) - len, ".%06d UTC\n", (int)tv.tv_usec);
     debug_write(line);
     debug_serial_ok("date");
     return 0;
